implement using bitmap ids and uninv ids

+++ include/actionmodern/runtime_api/BitmapData.h
@@ -9,7 +9,7 @@ typedef struct
 	u32 width;
 	u32 height;
 	
-	u32 char_id;
+	u16 char_id;
 } BitmapData;
 
 void BitmapData_new(SWFAppContext* app_context, ASObject* this, u32 num_args);
